Moved rabbit test to separate package.
